[Delete] SearchRecipeController
